moved vertex cover test classes to corresponding vertexcover test package
