+++ jgrapht-core/src/test/java/org/jgrapht/alg/vertexcover/VertexCoverTest.java
@@ -15,7 +15,7 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.alg;
+package org.jgrapht.alg.vertexcover;
 
 import java.util.*;
